+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Unreleased
 
+- Add Caddyfile directives for configuring BadgerDB. The in-memory mode is not default now.
 - Do not cache <200 or 400>= status response